requirements.txt
pillow
pyyaml
torch
torchvision
